CartographicRectangle
